Mostly successful refactor of detail-page
