+++ ui/src/app/cohort-review/detail-header/detail-header.component.spec.ts
@@ -1,0 +1,28 @@
+import {async, ComponentFixture, TestBed} from '@angular/core/testing';
+import {ClarityModule} from '@clr/angular';
+
+import {DetailHeaderComponent} from './detail-header.component';
+
+
+describe('DetailHeaderComponent', () => {
+  let component: DetailHeaderComponent;
+  let fixture: ComponentFixture<DetailHeaderComponent>;
+
+  beforeEach(async(() => {
+    TestBed.configureTestingModule({
+      imports: [ClarityModule],
+      declarations: [DetailHeaderComponent],
+    })
+    .compileComponents();
+  }));
+
+  beforeEach(() => {
+    fixture = TestBed.createComponent(DetailHeaderComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  xit('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});
