+++ ui/src/app/cohort-review/detail-page/detail-page.html
@@ -1,149 +1,11 @@
 <div class="detail-page" [class.sidebar-open]="sidebarOpen">
 
-  <div class="nav-header">
-    <button
-      type="button"
-      class="btn btn-sm"
-      title="Go Back to the Participant Table"
-      (click)="backToTable()">
-      <clr-icon shape="angle-double" dir="left"></clr-icon>
-      Table
-    </button>
-    <h4>Participant {{ participant.id }}</h4>
-    <div class="btn-group btn-sm">
-      <button
-        type="button"
-        class="btn"
-        title="Go To the Prior Participant"
-        [disabled]="isFirstParticipant"
-        (click)="previous()">
-        <clr-icon shape="caret" dir="left"></clr-icon>
-        Prev
-      </button>
-      <button
-        type="button"
-        class="btn"
-        title="Go To the Next Participant"
-        [disabled]="isLastParticipant"
-        (click)="next()">
-        Next
-        <clr-icon shape="caret" dir="right"></clr-icon>
-      </button>
-    </div>
-  </div>
+  <app-detail-header [participant]="participant">
+  </app-detail-header>
 
-  <clr-tabs>
-    <clr-tab>
-      <button clrTabLink id="detail-summary-tab">Summary</button>
-      <clr-tab-content id="detail-summary-content" *clrIfActive>
-        <em>Charts go here</em>
-      </clr-tab-content>
-    </clr-tab>
-
-    <clr-tab>
-      <button clrTabLink id="detail-conditions-tab">Conditions</button>
-      <clr-tab-content id="detail-conditions-content" *clrIfActive>
-
-        <clr-datagrid>
-          <clr-dg-column [clrDgField]="'description'">
-            Description
-          </clr-dg-column>
-          <clr-dg-column [clrDgField]="'standard'">
-            Standard Vocabulary (SNOMED)
-          </clr-dg-column>
-          <clr-dg-column [clrDgField]="'source'">
-            Source Vocabulary (ICD9CM, ICD10CM, HCPCS, SNOMED)
-          </clr-dg-column>
-          <clr-dg-column [clrDgField]="'value'">
-            Source Value
-          </clr-dg-column>
-          <clr-dg-column [clrDgField]="'date'">
-            Date
-          </clr-dg-column>
-          <clr-dg-column [clrDgField]="'ageAtEvent'">
-            Age at Event
-          </clr-dg-column>
-          <clr-dg-row *clrDgItems="let condition of DUMMY_CONDITIONS">
-            <clr-dg-cell>
-              {{ condition.description }}
-            </clr-dg-cell>
-            <clr-dg-cell>
-              {{ condition.standard }}
-            </clr-dg-cell>
-            <clr-dg-cell>
-              {{ condition.source }}
-            </clr-dg-cell>
-            <clr-dg-cell>
-              {{ condition.value }}
-            </clr-dg-cell>
-            <clr-dg-cell>
-              {{ condition.date | date:'yyyy-MM-dd HH:mm:ss' }}
-            </clr-dg-cell>
-            <clr-dg-cell>
-              {{ condition.ageAtEvent }}
-            </clr-dg-cell>
-          </clr-dg-row>
-        </clr-datagrid>
-
-      </clr-tab-content>
-    </clr-tab>
-
-    <clr-tab>
-      <button clrTabLink id="detail-procedures-tab">Procedures</button>
-      <clr-tab-content id="detail-procedures-content" *clrIfActive>
-        <em>Charts go here</em>
-      </clr-tab-content>
-    </clr-tab>
-
-    <clr-tab>
-      <button clrTabLink id="detail-drugs-tab">Drugs</button>
-      <clr-tab-content id="detail-drugs-content" *clrIfActive>
-        <em>Charts go here</em>
-      </clr-tab-content>
-    </clr-tab>
-
-    <clr-tab>
-      <button clrTabLink id="detail-measurements-tab">Measurements</button>
-      <clr-tab-content id="detail-measurements-content" *clrIfActive>
-        <em>Charts go here</em>
-      </clr-tab-content>
-    </clr-tab>
-
-    <clr-tab>
-      <button clrTabLink id="detail-observations-tab">Observations</button>
-      <clr-tab-content id="detail-observations-content" *clrIfActive>
-        <em>Charts go here</em>
-      </clr-tab-content>
-    </clr-tab>
-
-    <clr-tab>
-      <button clrTabLink id="detail-visits-tab">Visits</button>
-      <clr-tab-content id="detail-visits-content" *clrIfActive>
-        <em>Charts go here</em>
-      </clr-tab-content>
-    </clr-tab>
-
-    <clr-tab>
-      <button clrTabLink id="detail-physical-measurements-tab">Physical Measurements</button>
-      <clr-tab-content id="detail-physical-measurements-content" *clrIfActive>
-        <em>Charts go here</em>
-      </clr-tab-content>
-    </clr-tab>
-
-    <clr-tab>
-      <button clrTabLink id="detail-ppi-tab">PPI</button>
-      <clr-tab-content id="detail-ppi-content" *clrIfActive>
-        <em>Charts go here</em>
-      </clr-tab-content>
-    </clr-tab>
-
-    <clr-tab>
-      <button clrTabLink id="detail-device-tab">Device</button>
-      <clr-tab-content id="detail-device-content" *clrIfActive>
-        <em>Charts go here</em>
-      </clr-tab-content>
-    </clr-tab>
-  </clr-tabs>
+  <!-- Tab data stubbed in detail-tabs class -->
+  <app-detail-tabs>
+  </app-detail-tabs>
 </div>
 
 <div class="detail-sidebar">
