+++ ui/src/app/cohort-review/detail-tabs/detail-tabs.component.spec.ts
@@ -1,0 +1,28 @@
+import {async, ComponentFixture, TestBed} from '@angular/core/testing';
+import {ClarityModule} from '@clr/angular';
+
+import {DetailTabsComponent} from './detail-tabs.component';
+
+
+describe('DetailTabsComponent', () => {
+  let component: DetailTabsComponent;
+  let fixture: ComponentFixture<DetailTabsComponent>;
+
+  beforeEach(async(() => {
+    TestBed.configureTestingModule({
+      imports: [ClarityModule],
+      declarations: [ DetailTabsComponent ],
+    })
+    .compileComponents();
+  }));
+
+  beforeEach(() => {
+    fixture = TestBed.createComponent(DetailTabsComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});
