+++ ui/src/app/cohort-review/detail-tabs/detail-tabs.component.ts
@@ -1,0 +1,12 @@
+import {Component, OnInit} from '@angular/core';
+
+import {dummyData} from './dummy-data';
+
+@Component({
+  selector: 'app-detail-tabs',
+  templateUrl: './detail-tabs.component.html',
+  styleUrls: ['./detail-tabs.component.css']
+})
+export class DetailTabsComponent {
+  DUMMY_CONDITIONS = dummyData;
+}
