+++ ui/src/app/cohort-review/detail-tabs/dummy-data.ts
